+++ pbrocks/src/blocks/time-sidebar/index.js
@@ -1,105 +1,138 @@
+import { __ }  from '@wordpress/i18n';
 
-import { __ } from '@wordpress/i18n';
+import {
+    Fragment,
+} from '@wordpress/element';
 
 import { registerBlockType } from '@wordpress/blocks';
 
 import {
-    withState,
-    setState,
-} from '@wordpress/data';
+    __experimentalGetSettings,
+} from '@wordpress/date';
 
 import {
     PanelBody,
     PanelRow,
+    DateTimePicker,
 } from '@wordpress/components';
 
 import {
-	InspectorControls,
-	RichText,
+    InspectorControls,
 } from '@wordpress/block-editor';
 
-registerBlockType( 'gatherpress/time-sidebar', {
-    title: 'Gather: Times',
-    icon: { 
-        background: '#23c3aa',
+
+registerBlockType( 'gatherpress/event-times', {
+    title: __( 'Event Times', 'gatherpress' ),
+    icon: {
         foreground: '#fff',
+        background: '#23c3aa',
         src: 'clock',
     },
     category: 'gatherpress',
+    keywords: [
+        __( 'section', 'gatherpress'  ),
+        __( 'event', 'gatherpress'  ),
+    ],
     attributes: {
-        content: {
-            type: 'array',
-            source: 'children',
-            selector: 'p',
-        },
-        date: {
+        startDate: {
             type: 'string',
-        }
-    },
-    example: {
-        attributes: {
-            content: 'Hello World',
-            date: '31 July 2001',
+        },
+        endDate: {
+            type: 'string',
         },
     },
-    edit: ( props ) => {
-        const { attributes: { content, date }, setAttributes, className } = props;
-        const onChangeContent = ( newContent ) => {
-            setAttributes( { content: newContent } );
-        };
- 
-		const StartDateTimePicker = withState( {
-		    date: new Date(),
-		} )( ( { date, setState } ) => {
-		    const settings = __experimentalGetSettings();
-		 
-		    // To know if the current timezone is a 12 hour time with look for an "a" in the time format.
-		    // We also make sure this a is not escaped by a "/".
-		    const is12HourTime = /a(?!\\)/i.test(
-		        settings.formats.time
-		            .toLowerCase() // Test only the lower case a
-		            .replace( /\\\\/g, '' ) // Replace "//" with empty strings
-		            .split( '' ).reverse().join( '' ) // Reverse the string and test for "a" not followed by a slash
-		    );
-		 
-		    return (
-		        <DateTimePicker
-		            currentDate={ date }
-		            onChange={ ( date ) => setState( { date } ) }
-		            is12Hour={ is12HourTime }
-		        />
-		    );
-		} );
+    edit: function( {  attributes, setAttributes,className }) {
+
+        const settings = __experimentalGetSettings();
+        const is12HourTime = /a(?!\\)/i.test(
+            settings.formats.time
+                .toLowerCase() // Test only the lower case a
+                .replace( /\\\\/g, '' ) // Replace "//" with empty strings
+                .split( '' ).reverse().join( '' ) // Reverse the string and test for "a" not followed by a slash
+        );
+
+        const { 
+            startDate,
+            endDate
+        } = attributes;
 
         return (
-            <>
-            <InspectorControls>
-            <PanelBody
-                title="Time Input Sidebar"
-                icon="welcome-widgets-menus"
-                initialOpen={ true }
-            >
-            <PanelRow>
-                <h3>Time Inputs Panel and Labels</h3>
-            </PanelRow>
-                <PanelRow>
-                    <div className={className} >
-                        <StartDateTimePicker />
+            <Fragment>
+                <InspectorControls>
+                    <PanelBody
+                        title={ __( 'Event Start Time', 'gatherpress' ) }
+                        initialOpen={ true }
+                    >
+                    <PanelRow>
+                     <DateTimePicker
+                            currentDate={ startDate }
+                            onChange={ ( date ) => {
+                                setAttributes( {
+                                    startDate: date,
+                                } );
+                            } }
+                            locale={ settings.l10n.locale }
+                            is12Hour={ is12HourTime }
+                        />
+                    </PanelRow>
+                    </PanelBody>
+
+                    <PanelBody
+                        title={ __( 'Event End Time', 'gatherpress' ) }
+                        initialOpen={ false }
+                    >
+                    <PanelRow>
+                    <DateTimePicker
+                            currentDate={ endDate }
+                            onChange={ ( date ) => {
+                                setAttributes( {
+                                    endDate: date,
+                                } );
+                            } }
+                            locale={ settings.l10n.locale }
+                            is12Hour={ is12HourTime }
+                        />
+                    </PanelRow>
+
+                    </PanelBody>
+
+                </InspectorControls>
+                <section
+                    className={ className }
+                    style={{border:"1px solid"}}
+                >   
+                    <div>
+                    <h3>Start Time =  { startDate }</h3>
                     </div>
-                </PanelRow>
-                </PanelBody>
-            </InspectorControls>
-            <RichText
-                tagName="p"
-                className={ className }
-                onChange={ onChangeContent }
-                value={ content }
-            />
-            
-            </>
+                    <div>
+                    <h3>End Time = { endDate }</h3>
+                    </div>
+                </section>
+            </Fragment>
         );
     },
-    save: ( props ) => {
-        return <RichText.Content tagName="p" value={ props.attributes.content } />;
+
+    save: function( { attributes, className }) {
+        const { 
+            startDate,
+            endDate
+        } = attributes;
+
+            return (
+                <section
+                    className={ className }
+                    style={{border:"1px solid"}}
+                >   
+                <div>
+                <h3>Start Time =  { startDate }</h3>
+                </div>
+                <div>
+                <h3>End Time = { endDate }</h3>
+                </div>
+                </section>
+            );
+
+
     },
+
 } );
